requirements.txt
python-dotenv
telethon
cryptg
pyromod
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.23.zip
tgcrypto